# docs/mobile-api-contract/openapi-mobile-v1.yaml
openapi: 3.1.0
info:
  title: SpotLink Mobile API Contract
  version: 0.1.0-mobile-draft
  summary: Mobile-facing contract for the current SpotLink backend snapshot.
  description: >
    This draft mirrors the current unversioned backend paths under /api. A
    /api/v1 compatibility surface is now implemented for mobile-critical
    endpoints with the same controller logic and DTOs.
...
servers:
  - url: https://api.spotlink.app
    description: Production candidate
  - url: https://api-staging.spotlink.app
    description: Staging candidate
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Cookie/session login
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              login:
                value:
                  email: casey.customer@example.com
                  password: CorrectHorse123
      responses:
        "200":
          description: Authenticated cookie session
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/token:
    post:
      tags: [Auth]
      security: []
      summary: Mobile bearer token login
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileTokenRequest"
            examples:
              token:
                value:
                  email: casey.customer@example.com
                  password: CorrectHorse123
      responses:
        "200":
          description: Mobile access token
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileTokenResponse"
              examples:
                token:
                  value:
                    accessToken: sample-access-token
                    refreshToken: sl_refresh_example
                    tokenType: Bearer
                    expiresIn: 900
                    expiresInSeconds: 900
                    refreshExpiresInSeconds: 2592000
                    issuedAt: "2026-04-22T13:00:00Z"
                    expiresAt: "2026-04-22T13:15:00Z"
                    refreshExpiresAt: "2026-05-22T13:00:00Z"
                    user:
                      id: 11111111-1111-4111-8111-111111111111
                      email: casey.customer@example.com
                      firstName: Casey
                      lastName: Customer
                      phone: "+15551234567"
                      avatarUrl: null
                      bio: null
                      roles: [CUSTOMER]
                      operatorId: null
                      registrationStatus: ACTIVE
                      createdAt: "2026-04-22T12:00:00Z"
                    roles: [CUSTOMER]
        default:
          $ref: "#/components/responses/Error"
  /api/auth/token/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate mobile refresh token and issue new access token
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Rotated mobile token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileTokenResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Current recommended native iOS auth model.
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Current browser/session auth model.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: Required for cookie/session state-changing browser requests.
  parameters:
    RequestIdHeader:
      in: header
      name: X-Request-Id
      required: false
      schema:
        type: string
      description: Client-generated correlation ID.
...
  headers:
    RequestId:
      description: Server correlation ID.
      schema:
        type: string
...
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            validation:
              $ref: "#/components/examples/ValidationError"
            authLockout:
              $ref: "#/components/examples/AuthLockoutError"
  examples:
...
    ValidationError:
      value:
        status: 400
        code: VALIDATION_ERROR
        message: Input validation failed
        requestId: mob-11111111-1111-4111-8111-111111111111
        details:
          email: must be a well-formed email address
        timestamp: "2026-04-22T13:00:00Z"
        path: /api/auth/token
    AuthLockoutError:
      value:
        status: 423
        code: AUTH_TEMPORARILY_LOCKED
        message: Too many failed sign-in attempts. Try again later.
        requestId: mob-lockout-001
        details:
          retryAfterSeconds: "900"
        timestamp: "2026-06-06T13:00:00Z"
        path: /api/auth/token
  schemas:
    UUID:
      type: string
      format: uuid
    Instant:
      type: string
      format: date-time
...
    ApiError:
      type: object
      required: [status, code, message, requestId, timestamp, path]
      properties:
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        requestId:
          type: [string, "null"]
        details:
          type: [object, "null"]
          additionalProperties: true
        timestamp:
          $ref: "#/components/schemas/Instant"
        path:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    MobileTokenRequest:
      $ref: "#/components/schemas/LoginRequest"
    MobileTokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, expiresInSeconds, refreshExpiresInSeconds, issuedAt, expiresAt, refreshExpiresAt, user, roles]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
          deprecated: true
          description: Backward-compatible alias of expiresInSeconds.
        expiresInSeconds:
          type: integer
          format: int64
        refreshExpiresInSeconds:
          type: integer
          format: int64
        tokenType:
          type: string
          const: Bearer
        issuedAt:
          $ref: "#/components/schemas/Instant"
        expiresAt:
          $ref: "#/components/schemas/Instant"
        refreshExpiresAt:
          $ref: "#/components/schemas/Instant"
        user:
          $ref: "#/components/schemas/UserProfile"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRole"
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
        deviceId:
          type: [string, "null"]
          maxLength: 160
...
    AuthResponse:
      type: object
      required: [authenticated, user, message]
      properties:
        authenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/UserProfile"
        message:
          type: string
...
    UserRole:
      type: string
      enum: [CUSTOMER, OPERATOR, SUPPORT, ADMIN]
    RegistrationStatus:
      type: string
      enum: [INCOMPLETE, ACTIVE, SUSPENDED, DELETED]
...
    UserProfile:
      type: object
      required: [id, email, firstName, lastName, roles, registrationStatus, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: [string, "null"]
        avatarUrl:
          type: [string, "null"]
          format: uri
        bio:
          type: [string, "null"]
        roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRole"
        operatorId:
          type: [string, "null"]
          format: uuid
        registrationStatus:
          $ref: "#/components/schemas/RegistrationStatus"
        createdAt:
          $ref: "#/components/schemas/Instant"
...